requirements.txt
flask
flask-cors
nbconvert
nbformat
pyyaml
notebook
requests
python-dotenv
pandas
seaborn
scikit-learn
progress
newspaper3k
wikipedia
emoji
lxml_html_clean